// WebSite/src/models/models.ts
export class BaseModel {
  id: number = 0;
  creationUserId?: number;
  creationDate?: Date;
  modificationUserId?: number;
  modificationDate?: Date;
}

export class ResponseResult<T> {
  data: T;
  status: StatusType = 0;
  errors: string[] = [];
  // added code to get the data from the data base per page size
  totalRecords!: number;
  // added code to get the data from the data base per page size
  pageNumber!: number;    // current page number
  // added code to get the data from the data base per page size
  pageSize!: number;      // page size returned
  // added code to get the data from the data base per page size
  // optionally:
  succeeded: boolean = true;
  constructor(_data: T) {
    this.data = _data;
  }
}

export class LoginResonse {
  id: number = 0;
  name: string = '';
  type: UserType = 1;
  token: string = '';
}

export enum StatusType {
  Failed = 0,
  Success = 1,
}
// ...
export class Payment extends BaseModel {
  clientName: string = '';
  paymentId: number = 0;
  paymentType: number = 0;
  autoId: number = 0;
  vinFormNumber: string = '';
  categoryId: number = 0;
  amount: number = 0;
  payDate?: Date;
  paymentMethod: number = 0;
  notes: string = '';
}

export class PaymentDetails {
  id: number = 0;
  cashType: number = 0;
  buyType: number = 0;
  notes: string = '';
  payDate?: Date;
  amount: number = 0;
  paymentId: number = 0;
}
// ...
export class SearchPayment {
  isSearch: boolean = false;
  autoId: number = 0;
  vinNo: string = '';
  purchaseDate!: Date;
  clientId: number = 0;
}